requirements_dev.txt
# ASEED Development Requirements
# Dodatkowe pakiety dla rozwoju, testowania i analizy

# Core requirements (from requirements.txt)
pyspark==3.5.0
kafka-python==2.0.2
flask==2.3.3
flask-socketio==5.3.6
requests==2.31.0
pandas==2.1.3

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
unittest-xml-reporting==3.2.0

# Data Science and Visualization
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
numpy==1.24.4

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
myst-parser==2.0.0

# Performance Monitoring
psutil==5.9.6
memory-profiler==0.61.0

# Enhanced Order Simulation  
faker==20.1.0

# Optional: Kafka Management UI (if needed)
# kafka-manager

# Optional: Spark Web UI (built-in)
# Access at: http://localhost:4040 during Spark job execution

# Development Environment
python-dotenv==1.0.0
ipython==8.17.2

# Containerization (if needed)
# docker==6.1.3
# docker-compose==1.29.2
